% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Cluster all interesting (differentially expressed) genes using hierarchical clustering (based on all groups together).}
\usage{
clustering(resultAOV, clust.metric = NULL, clust.method = NULL,
  dist.matrix = NULL)
}
\arguments{
\item{resultAOV}{A list of data frames of gene expressions (genes in rows, samples in columns).}

\item{clust.metric}{The method to calculate a distance measure used in hierarchical clustering, possible names: \code{"euclidean"},
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.}

\item{clust.method}{The agglomeration method of clustering genes. This should be one of: \code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average" (= UPGMA)}, \code{"mcquitty" (= WPGMA)}, \code{"median" (= WPGMC)} or \code{"centroid" (= UPGMC)}.}

\item{dist.matrix}{A matrix with calculated distances to be used as a metric by \code{hclust} function. The default distance is \code{euclidean}.}
}
\value{
An object of class \code{hclust}.
}
\description{
Cluster all interesting (differentially expressed) genes using hierarchical clustering (based on all groups together).
}
\examples{
clustering(exanova, clust.metric = 'manhattan', clust.method = 'average')

}

