% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exanova}
\alias{exanova}
\title{Expressions of differently expressed human genes.}
\format{A list of 3 data frames with 50 rows each and 173, 190, and 122 variables respectively. Each variable corresponds to one probe.}
\usage{
exanova
}
\description{
A dataset containing gene expressions of 50 human genes returned by function
\code{aovTopTest(exrtcga, top = 50, sig.level = 0.1)}.
}
\keyword{datasets}

