% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exrtcga}
\alias{exrtcga}
\title{Expressions of human genes.}
\format{A list of 3 data frames with 300 rows each and 173, 190, and 122
 variables respectively.
 Each variable corresponds to one probe.}
\usage{
exrtcga
}
\value{
data
}
\description{
A dataset containing gene expressions of 300 human genes
randomly chosen from a list returned by function
\code{prepareData(RTCGA = TRUE, cohorts = c('leukemia', 'colon', 'bladder'))}.
}
\keyword{datasets}

