% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendGO.R
\name{extendGO}
\alias{extendGO}
\title{Find the description of characteristic GO functions.}
\usage{
extendGO(TopGOs)
}
\arguments{
\item{TopGOs}{A list of vectors with significant GO IDs.}
}
\value{
A data frame with an extended description
 (term, definition, and ontology) of gene ontology terms.
}
\description{
Find the description of characteristic GO functions.
}
\examples{
tukey.results <- groupByTukey(exanova, sig.level = 0.1)
# find all and then top GO terms
# geneUniverse is a universe of all genes
# of interest, i.e. all genes measured by microarray
geneUniverse <- rownames(exrtcga[[1]])
all.gos <- findAllGOs(geneUniverse, tukey.results, min = 8, max = 400)
top.gos <- findTopGOs(all.gos, sig.level = 0.1, top = 2)
extendGO(top.gos)

}

