% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllGOs.R
\name{findAllGOs}
\alias{findAllGOs}
\title{Find all ontology terms annotated with genes.}
\usage{
findAllGOs(geneUni, geneSel, onto = c("MF", "BP", "CC"), min = 5,
  max = 500, parallel = FALSE)
}
\arguments{
\item{geneUni}{A vector of names of all genes that
were primarily included in the analysis,
i.e. all genes measured by the microarray.}

\item{geneSel}{A list of groups of selected genes
or an hclust object representing clustered genes.}

\item{onto}{An ontology or ontologies to be searched for
significant GO terms, at least one of 'MF' (molecular function),
'BP' (biological process), 'CC' (cellular component).}

\item{min}{A minimum number of genes that a function needs
to be represented by to be considered in the analysis.}

\item{max}{A maximum number of genes that a function needs
to be represented by to be considered in the analysis.}

\item{parallel}{A logical value indicating if a task
should be run on more than one core.
The default is \code{FALSE}.}
}
\value{
A list of lists with all gene ontology terms
 identifiers relevant to each group of genes and p-values for Fisher's test.
}
\description{
Find all ontology terms annotated with genes.
}
\examples{
tukey.results <- groupByTukey(exanova, 0.1)
#geneUniverse is a universe of all genes of interest
geneUniverse <- rownames(exrtcga[[1]])
findAllGOs(geneUniverse, tukey.results)

}

