% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgenecluster.R
\name{geneclusterplot}
\alias{geneclusterplot}
\title{Constructor function for the class \code{geneclusterplot}.}
\usage{
geneclusterplot(printout, genecl, TopGOs)
}
\arguments{
\item{printout}{A data frame with labels, genes' names
and GO IDs returned by \code{print.GOcluster} function.}

\item{genecl}{An object of class \code{genecluster}.}

\item{TopGOs}{A list of vectors with significant GO IDs.}
}
\value{
An object of class \code{geneclusterplot}.
}
\description{
Constructor function for the class \code{geneclusterplot}.
}
\examples{
clustered <- clustering(exanova,
 clust.metric = 'manhattan', clust.method = 'average')
genecl <- unbundleCluster(clustered)
# find all and then top GO terms
# geneUniverse is a universe of all genes of interest
# i.e. all genes measured by microarray
geneUniverse <- rownames(exrtcga[[1]])
all.gos <- findAllGOs(geneUniverse, clustered)
top.gos <- findTopGOs(all.gos, sig.level = 0.1, top = 4)
# create an object of class \\code{GO}
printout <- GO(all.gos, top.gos)
geneclusterplot(printout, genecl, top.gos)

}

