% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOpro.R
\docType{package}
\name{GOpro}
\alias{GOpro}
\alias{GOpro-package}
\title{GOpro: find the most characteristic gene ontology terms for groups of genes}
\description{
Finds the most characteristic gene ontology terms for groups of genes
}

