% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByProfiles.R
\name{groupByProfiles}
\alias{groupByProfiles}
\title{Genes grouped by profiles determined by the result returned from \code{tukeyHSDTest} function.}
\usage{
groupByProfiles(expr, tukey.results, sig.level)
}
\arguments{
\item{expr}{A list of data frames obtained from \code{aovTopTest}.}

\item{tukey.results}{A data frame returned by \code{TukeyHSDTest}.}

\item{sig.level}{A numeric value, a significance level used in Tukey's all pairwise comparisons (\code{groupByTukey}).}
}
\value{
A list of groups of genes named by profiles.
}
\description{
Genes grouped by profiles determined by the result returned from \code{tukeyHSDTest} function.
}

