% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByTukey.R
\name{groupByTukey}
\alias{groupByTukey}
\title{Determine group of genes by all pairwise comparisons (Tukey's method).}
\usage{
groupByTukey(expr, parallel = FALSE, sig.level = 0.05)
}
\arguments{
\item{expr}{A named list of data frames of significantly different gene expressions with genes in rows and samples in columns.
rows should be named with genes' aliases. The order of genes has to be the same for each data frame. The names of list elements should refer to groups of different cases.}

\item{parallel}{A logical value indicating if a task should be run on more than one core. The default value is \code{FALSE}.}

\item{sig.level}{A numeric value, a significance level used in BH correction for multiple testing. The default value is 0.05.}
}
\value{
A data frame with groups of genes named by profiles. Profiles are determined by the differences between groups.
}
\description{
Determine group of genes by all pairwise comparisons (Tukey's method).
}
\examples{
groupByTukey(exanova, sig.level = 0.1)

}

