% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByProfiles.R
\name{makeNames}
\alias{makeNames}
\title{Create profiles names of Tukey test results.}
\usage{
makeNames(part.tukey.res, sig.level, n)
}
\arguments{
\item{part.tukey.res}{A vector of a list of
a result returned by \code{TukeyHSDTest}.}

\item{sig.level}{A numeric significance level of a test.}

\item{n}{A number of compared groups.}
}
\value{
A character vector of a name denoting
results of all pairwise Tukey comparisons.
}
\description{
Create profiles names of Tukey test results.
}

