% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontologyOver.R
\name{ontologyOver}
\alias{ontologyOver}
\title{Calculate frequencies of each ontology in each group (CC, BP, MF).}
\usage{
ontologyOver(extendedGO)
}
\arguments{
\item{extendedGO}{A data frame with a one column of group and another column with GO term (returned by \code{extendGO}).}
}
\value{
A data frame with the most frequently occurring ontologies in each group.
}
\description{
Calculate frequencies of each ontology in each group (CC, BP, MF).
}
\examples{
tukey.results <- groupByTukey(exanova, sig.level = 0.1)
# find all and then top GO terms
# geneUniverse is a universe of all genes of interest, i.e. all genes measured by microarray
geneUniverse <- rownames(exrtcga[[1]])
all.gos <- findAllGOs(geneUniverse, tukey.results)
top.gos <- findTopGOs(all.gos, sig.level = 0.1, top = 3)
extended <- extendGO(top.gos)
ontologyOver(extended)

}

