% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgenecluster.R
\name{plotg}
\alias{plotg}
\title{Plot gene dendrogram with labels.}
\usage{
plotg(x, over.represented)
}
\arguments{
\item{x}{an object of class \code{geneclusterplot}.}

\item{over.represented}{a logical value, if \code{TRUE} then
over represented GO categories are coded with colors for each branch.}
}
\value{
n
}
\description{
Plot gene dendrogram with labels.
}
\examples{
clustered <- clustering(exanova,
 clust.metric = 'manhattan', clust.method = 'average')
genecl <- unbundleCluster(clustered)
# find all and then top GO terms
# geneUniverse is a universe of all genes of interest
geneUniverse <- rownames(exrtcga[[1]])
all.gos <- findAllGOs(geneUniverse, clustered)
top.gos <- findTopGOs(all.gos, sig.level = 0.05, top = 3)
# create an object of class \\code{GO}
printout <- GO(all.gos, top.gos)
plotg(geneclusterplot(printout, genecl, top.gos),
 over.represented = TRUE)

}

