% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Rearrange data and select all genes with complete observation in order to analyze it using GOpro package.
Prepare an input dataset or use \code{RTCGA.PANCAN12} data.}
\usage{
prepareData(RTCGA = FALSE, cohorts = c("colon", "bladder", "leukemia"),
  data = NULL)
}
\arguments{
\item{RTCGA}{logical, if TRUE the data is taken from the RTCGA.PANCAN12 package, otherwise provide data using parameter /textit{data} as a vector of data frames. Default value is FALSE.}

\item{cohorts}{a vector of names of cohorts to be derived from the RTCGA.PANCAN12 package if RTCGA is TRUE, at least two names of: 'colon', 'ovarian', bladder', 'leukemia', 'glioblastoma', 'lung adenocarcinoma', 'BRCA',
lung squamous carcinoma', 'kidney', 'endometrioid', 'rectal', 'head and neck', or a vector of names of cohorts provided by data parameter.}

\item{data}{a list of data frames each with rows named with genes' aliases. The number of a column in each data frame should correspond to the same case.}
}
\value{
A list of data frames of genes expressions or other gene activity related measure with rows named as genes' aliases.
}
\description{
Rearrange data and select all genes with complete observation in order to analyze it using GOpro package.
Prepare an input dataset or use \code{RTCGA.PANCAN12} data.
}
\examples{
prepareData(RTCGA = TRUE, cohorts = c('BRCA', 'ovarian', 'colon'))

}

