% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printgenecluster.R
\name{print.genecluster}
\alias{print.genecluster}
\title{Print groups of clustered genes.}
\usage{
\method{print}{genecluster}(x, ...)
}
\arguments{
\item{x}{An object of class \code{genecluster}.}

\item{...}{Additional arguments to be passed (not in use).}
}
\description{
Print groups of clustered genes.
}
\examples{
unbundleCluster(clustering(exanova, clust.metric = 'manhattan', clust.method = 'average'))

}

