% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{print.GO}
\alias{print.GO}
\title{Print grouped genes with the information of the labels and GO term's IDs.}
\usage{
\method{print}{GO}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GO}.}

\item{...}{Additional arguments to be passed (not in use).}
}
\value{
A data frame with labels, gene's aliases and GO terms for each group.
}
\description{
Print grouped genes with the information of the labels and GO term's IDs.
}
\examples{
clustered <- clustering(exanova)
genecl <- unbundleCluster(clustered)
# find all and then top GO terms
# geneUniverse is a universe of all genes of interest
geneUniverse <- rownames(exrtcga[[1]])
all.gos <- findAllGOs(geneUniverse, clustered, max = 300, onto = c('MF', 'BP'))
top.gos <- findTopGOs(all.gos, sig.level = 0.05, top = 3)
# create an object of class \\code{GO}
printout <- GO(all.gos, top.gos)

}

