% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyHSDTest.R
\name{tukeyHSDTest}
\alias{tukeyHSDTest}
\title{Call the calculation of p-values across all groups of genes
using Tukey's Honest Distance test optionally on more than one core (actually \code{#(available cores) - 1}).}
\usage{
tukeyHSDTest(expr, parallel = FALSE)
}
\arguments{
\item{expr}{A list of data frames of gene expressions with genes in rows and samples in columns.
the names of rows represent genes' aliases. The order of genes has to be the same for each data frame.}

\item{parallel}{A logical value indicating if a task should be run on more than one core. The default is \code{FALSE}.}
}
\value{
A list of \code{c(mean.val, p.val)}.
}
\description{
Call the calculation of p-values across all groups of genes
using Tukey's Honest Distance test optionally on more than one core (actually \code{#(available cores) - 1}).
}

