% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbundleCluster.R
\name{unbundleCluster}
\alias{unbundleCluster}
\title{Find all groups - rearrange the dendrogram: find which elements were merged in each step.}
\usage{
unbundleCluster(hc)
}
\arguments{
\item{hc}{An object of class hclust.}
}
\value{
An object of class \code{genecluster}, a list of labels of elements that were merged in each step.
}
\description{
Find all groups - rearrange the dendrogram: find which elements were merged in each step.
}
\examples{
clusteredGenes <- clustering(exanova, clust.metric = 'manhattan', clust.method = 'average')
unbundleCluster(clusteredGenes)

}

