% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exrtcga}
\alias{exrtcga}
\title{Expressions of human genes.}
\format{A MultiAssayExperiment object of 3 listed
experiments with user-defined names and respective classes.
Containing an ExperimentList class object of length 3:
[1] leukemia: matrix with 300 rows and 173 columns
[2] colon: matrix with 300 rows and 190 columns
[3] bladder: matrix with 300 rows and 122 columns.}
\usage{
exrtcga
}
\value{
data
}
\description{
A dataset containing gene expressions of 300 human genes
randomly chosen from a list returned by function
\code{prepareData(RTCGA = TRUE, cohorts = c('leukemia', 'colon', 'bladder'))}.
}
\keyword{datasets}

